Assignment completed
Completed parallel-add-list and ran on Palmetto with 2, 4, 8, 16 and 32 processors.

To run on Palmetto you must have a copy of the parallel-add-list executable in the root directory (cd ~). Then run `make pbs` and it'll run for 2, 4, 8, 16 and 32 processors and then output .o###### files with the result.

Also able to run it locally using `mpirun -n [#] ./parallel-add-list`

+++ Assignment 1/MyMPI.h
@@ -47,7 +47,7 @@ void read_col_striped_matrix (char *, void ***, void **,
 void read_row_striped_matrix (char *, void ***, void **,
         MPI_Datatype, int *, int *, MPI_Comm);
 void read_block_vector (char *, void **, MPI_Datatype,
-        int *, MPI_Comm);
+        long long int *, MPI_Comm);
 void read_replicated_vector (char *, void **, MPI_Datatype,
         int *, MPI_Comm);
